move background notifications flutter logic to separate file
